requirements.txt
africastalking==1.2.7
ansible==9.2.0
ansible-core==2.16.3
apturl==0.5.2
asgiref==3.7.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
cli-helpers==2.3.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.8
contextlib2==21.6.0
cryptography==42.0.2
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.8.0rc2
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.2
django-cors-headers==4.2.0
django-nose==1.4.7
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
django-tenants==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
duplicity==0.8.21
fasteners==0.14.1
frappe-bench==5.22.3
future==0.18.2
git-review==2.2.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==1.1.2
honcho==1.1.0
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
inflection==0.5.1
install==1.3.5
isort==5.13.2
jeepney==0.7.1
Jinja2==3.1.3
jwcrypto==1.5.3
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.2
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
netifaces==0.11.0
nose==1.3.7
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.9.3
pendulum==3.0.0
pexpect==4.8.0
pgcli==4.0.1
pgspecial==2.1.1
phonenumbers==8.13.35
Pillow==10.1.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psycopg==3.1.17
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-crontab==2.6.0
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python3-openid==3.2.0
pytz==2024.1
pyxdg==0.27
PyYAML==6.0.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.4.0
resolvelib==1.0.1
schema==0.7.5
SecretStorage==3.3.1
semantic-version==2.8.5
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
systemd-python==234
tabulate==0.9.0
time-machine==2.13.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.0
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.13
xdg==5
xkit==0.0.0
zipp==1.0.0
